User Logins added
models.py amended to generate and check password hash using
werkzeug.security.

__init__.py amended:
- flask-login extension added to venv and its LoginManager object is
instantiated and registered with the app in __init__.py
- LoginManager(app) added
- login.login_view = 'login' set up

Flask-Login requires us to implement 4 properties and methods to work
namely, is_authenticated, is_active, is_anonymous, and get_id().
Flask-login provides a mixin class called UserMixin that includes
generic implementations of these for most user model classes. So we just
add the UserMixin as a parent of the User class in the models.py to
inherit from that.

Flask-login needs a user_loader function to work. so we create the
load_user(id) and decorate it with @login.user_loader in models.py.

Amended routes.py:
- login view function to make use of actual users.
- added logout view function, using logout_user() from flask_login
- @login_required decorator added to routes index() so if not logged in
the user will be automatically redirected to the login page.
- If logged in, the user will be redirected to either the index page or
whatever page he attempted to visit before he logged in. This was stored
in the request.args.get('next'). If an attacker inserted an url to a
malicious site in the next argument, we redirect to the index page.
- No need to send the user to the index page template as it uses the
current user now.
- Added /register route and its view function (both get/post requests)
to save the user to the database and redirect user to login page

forms.py amended:
- Class RegistrationForm(FlaskForm) created, including all validators

Template changes:

Amended base.html:
- conditionally switch login / logout link in the navigation bar

index.html:
- fake user replaced with current user

register.html added:
- New form with 4 fields (username, email, password 1 and 2 and submit button

login.html amended to add link to the register page.