Error Handling Added
Custom Error pages added
- errors.py adds error handlers for not_found_error (404) and
internal_error (500). We import errors.py from __init__.py and this
causes it to run.
- 404.html: shows File not found and gives link to index page.
- 500.html: shows unexpected error and gives link to index page.

Config.py amended:
- to add config for Email server

__init__.py amended:
- to add an email handler and a RotatingFileHandler
     to the app.logger when the app is not running in debug mode

Amended forms.py to fix the duplicate username bug by saving the
original username in an attribute of the form object in its constructor
and adding a custom validation method that uses it.

Amended routes.py to pass the current_user.username to the
EditProfileForm constructor.

Also learnt:
- how to run flask in debug mode: set FLASK_MODE=1